+++ src/main/scala/org/graphsense/storage/CassandraStorage.scala
@@ -8,7 +8,10 @@ import org.graphsense.Util._
 import scala.reflect.ClassTag
 import com.datastax.spark.connector.cql.CassandraConnector
 
-class CassandraStorage(spark: SparkSession) {
+class CassandraStorage(
+    spark: SparkSession,
+    bulkWriter: Option[SidecarBulkWriter] = None
+) {
 
   import spark.implicits._
   import com.datastax.spark.connector._
@@ -38,6 +41,25 @@ class CassandraStorage(spark: SparkSession) {
       table.select(columns: _*).toDS().as[T]
   }
 
+  /** Column names of a Cassandra table, read from system_schema. Used to
+    * project a DataFrame onto the table columns before a bulk write.
+    */
+  def tableColumnNames(keyspace: String, tableName: String): Seq[String] = {
+    import scala.collection.JavaConverters._
+    CassandraConnector(spark.sparkContext).withSessionDo { session =>
+      session
+        .execute(
+          "select column_name from system_schema.columns " +
+            s"where keyspace_name = '${keyspace}' " +
+            s"and table_name = '${tableName}';"
+        )
+        .all()
+        .asScala
+        .map(_.getString("column_name"))
+        .toSeq
+    }
+  }
+
   def store[T <: Product: RowWriterFactory](
       keyspace: String,
       tableName: String,
@@ -48,7 +70,16 @@ class CassandraStorage(spark: SparkSession) {
       s"Writing table ${tableName} - empty: ${isEmpty}"
     )
     time(s"Writing table ${tableName} - empty: ${isEmpty}") {
-      df.rdd.saveToCassandra(keyspace, tableName)
+      bulkWriter match {
+        case Some(writer) =>
+          writer.write(
+            keyspace,
+            tableName,
+            df.toDF,
+            tableColumnNames(keyspace, tableName)
+          )
+        case None => df.rdd.saveToCassandra(keyspace, tableName)
+      }
     }
   }
 }